divide Main.swift into smaller files
